feat: add menu to home page
