+++ app/layout.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import "./globals.css";
+import Menu from "../components/menu";
 
 export const metadata = {
   title: {
@@ -16,7 +17,12 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en">
-      <body className="text-dark">{children}</body>
+      <body className="text-dark">
+        <Menu />
+        <main className="flex flex-col justify-start items-center p-4 md:p-12 py-12 h-56 min-h-screen">
+          {children}
+        </main>
+      </body>
     </html>
   );
 }
