+++ components/menu/__tests__/menu.test.tsx
@@ -4,11 +4,17 @@ import userEvent from "@testing-library/user-event";
 
 import Menu from "../index";
 
+jest.mock("next/navigation", () => ({
+  usePathname() {
+    return "/me-and-my-site";
+  },
+}));
+
 describe("Menu", () => {
   test("renders the Menu with active URI /me-and-my-site", () => {
-    render(<Menu currentUri={"/me-and-my-site"} />);
+    render(<Menu />);
 
-    const activeMenuLink = screen.getByText("Me and my site")
+    const activeMenuLink = screen.getByText("Me and my site");
     const inactiveMenuLink = screen.getByText("Quality topics");
 
     expect(activeMenuLink).toHaveClass("border-primary");
@@ -16,13 +22,13 @@ describe("Menu", () => {
   });
 
   test("link starts hidden", () => {
-    render(<Menu currentUri={"/me-and-my-site"} />);
+    render(<Menu />);
 
     expect(screen.queryAllByRole("list")).not.toBeInTheDocument;
   });
 
   test("burger menu reveals links", async () => {
-    render(<Menu currentUri={"/me-and-my-site"} />);
+    render(<Menu />);
 
     const burger = screen.getByRole("button");
     userEvent.click(burger);
