--- components/menu/types.ts
@@ -1,3 +1,0 @@
-export type MenuProps = {
-  currentUri: string;
-};
